pyproject.toml
[project]
name = "classifier-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.34",
    "tensorflow>=2.20.0",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.41.0",
]
